refactor(mime): consolidate MIME utilities into epicenter package
Unified duplicate MIME type implementations from whispering and epicenter into a single module at `@epicenter/hq/mime`. The API follows the `mime` npm package convention with `getType(pathOrExtension)` and `getExtension(mimeType)`.

Key changes:
- Created `packages/epicenter/src/mime/index.ts` with ~60 MIME mappings
- Added `./mime` subpath export to epicenter
- Removed `apps/whispering/src/lib/constants/mime.ts` facade
- Co-located Tauri-specific `getMimeTypeFromPath` in `fs/desktop.ts`
- Added `@epicenter/hq` as whispering dependency
