+++ apps/whispering/package.json
@@ -43,6 +43,7 @@
 		"@anthropic-ai/sdk": "^0.55.0",
 		"@aptabase/tauri": "^0.4.1",
 		"@aptabase/web": "^0.4.3",
+		"@epicenter/hq": "workspace:*",
 		"@google/generative-ai": "^0.24.1",
 		"@mistralai/mistralai": "^1.9.18",
 		"@repo/constants": "workspace:*",
