+++ packages/epicenter/package.json
@@ -6,6 +6,7 @@
 	"exports": {
 		".": "./src/index.ts",
 		"./cli": "./src/cli/index.ts",
+		"./mime": "./src/mime/index.ts",
 		"./providers": "./src/providers/index.ts",
 		"./indexes/markdown": "./src/indexes/markdown/index.ts",
 		"./indexes/sqlite": "./src/indexes/sqlite/index.ts"
