+++ packages/epicenter/src/core/blobs/utils.ts
@@ -58,86 +58,3 @@ export function validateFilename(filename: string) {
 	return Ok(undefined);
 }
 
-/**
- * Common MIME types for blob storage.
- * Falls back to 'application/octet-stream' for unknown extensions.
- */
-const MIME_TYPES: Record<string, string> = {
-	// Images
-	png: 'image/png',
-	jpg: 'image/jpeg',
-	jpeg: 'image/jpeg',
-	gif: 'image/gif',
-	webp: 'image/webp',
-	svg: 'image/svg+xml',
-	ico: 'image/x-icon',
-	bmp: 'image/bmp',
-	tiff: 'image/tiff',
-	tif: 'image/tiff',
-
-	// Documents
-	pdf: 'application/pdf',
-	doc: 'application/msword',
-	docx: 'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
-	xls: 'application/vnd.ms-excel',
-	xlsx: 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
-	ppt: 'application/vnd.ms-powerpoint',
-	pptx: 'application/vnd.openxmlformats-officedocument.presentationml.presentation',
-
-	// Text
-	txt: 'text/plain',
-	md: 'text/markdown',
-	html: 'text/html',
-	htm: 'text/html',
-	css: 'text/css',
-	csv: 'text/csv',
-	xml: 'text/xml',
-
-	// Code
-	js: 'application/javascript',
-	mjs: 'application/javascript',
-	ts: 'text/typescript',
-	json: 'application/json',
-	yaml: 'text/yaml',
-	yml: 'text/yaml',
-
-	// Audio
-	mp3: 'audio/mpeg',
-	wav: 'audio/wav',
-	ogg: 'audio/ogg',
-	m4a: 'audio/mp4',
-	flac: 'audio/flac',
-	aac: 'audio/aac',
-	webm: 'audio/webm',
-
-	// Video
-	mp4: 'video/mp4',
-	avi: 'video/x-msvideo',
-	mov: 'video/quicktime',
-	mkv: 'video/x-matroska',
-
-	// Archives
-	zip: 'application/zip',
-	gz: 'application/gzip',
-	tar: 'application/x-tar',
-	rar: 'application/vnd.rar',
-	'7z': 'application/x-7z-compressed',
-
-	// Fonts
-	woff: 'font/woff',
-	woff2: 'font/woff2',
-	ttf: 'font/ttf',
-	otf: 'font/otf',
-	eot: 'application/vnd.ms-fontobject',
-};
-
-/**
- * Get the MIME type for a filename based on its extension.
- *
- * @param filename The filename to get the MIME type for
- * @returns The MIME type, or 'application/octet-stream' if unknown
- */
-export function getMimeType(filename: string): string {
-	const ext = filename.split('.').pop()?.toLowerCase();
-	return MIME_TYPES[ext ?? ''] ?? 'application/octet-stream';
-}
